Edit, update UI -> done 23/4
