+++ src/core/Bricks.ts
@@ -1,19 +1,20 @@
 import { Board } from './Board';
 import Game from './GameControler';
+import GameView from './GameView';
 export class Brick {
     private game: Game;
     private board: Board;
     private isCurrentBrickLanded: boolean;
-
-    id: number;
-    layout: any;
-    activeIndex: number;
-    colPos: number;
-    rowPos: number;
-    nextcolPos: number;
-    nextrowPos: number;
-    isLanded: boolean;
-    gameOver: any;
+    private gameView: GameView;
+    public id: number;
+    public layout: any;
+    public activeIndex: number;
+    public colPos: number;
+    public rowPos: number;
+    public nextcolPos: number;
+    public nextrowPos: number;
+    public isLanded: boolean;
+    public gameOver: any;
 
 
     constructor(id: number, game: Game) {
@@ -21,18 +22,14 @@ export class Brick {
         this.game = game;
         this.layout = this.game.getBrickLayout()[id];
         this.activeIndex = 0;
-
         this.colPos = 4;
         this.rowPos = -1;
         this.nextcolPos = 1;
         this.nextrowPos = 1;
-
         this.board = this.game.getBoard();
         this.isLanded = false;
         this.isCurrentBrickLanded = false;
         this.gameOver = false;
-
-
     }
 
     draw() {
@@ -55,8 +52,6 @@ export class Brick {
         }
     }
 
-
-
     clear() {
         for (let row = 0; row < this.layout[this.activeIndex].length; row++) {
             for (let col = 0; col < this.layout[this.activeIndex][row].length; col++) {
@@ -66,8 +61,6 @@ export class Brick {
             }
         }
     }
-
-
 
     clearNextBrick() {
         for (let row = 0; row < this.layout[this.activeIndex].length; row++) {
@@ -119,7 +112,6 @@ export class Brick {
         if (this.isLanded) {
             return;
         }
-
         if (!this.checkCollision(this.rowPos + 1, this.colPos, this.layout[this.activeIndex])) {
             this.clear();
             this.rowPos++;
@@ -128,22 +120,15 @@ export class Brick {
         } else {
             this.handleLanded();
             this.isCurrentBrickLanded = true;
-
-            // if (!this.board.gameOver) {
             this.game.generateNewBrick();
-            // }
-
         }
-
-
     }
 
 
 
     fixPosition(nextRow: number, nextCol: number, nextLayout: any) {
-        this.rowPos = nextRow - 1; // Move the brick to the next position
+        this.rowPos = nextRow - 1;
         this.draw();
-
         if (!this.checkCollision(this.rowPos + 1, this.colPos, this.layout[this.activeIndex])) {
             this.clear();
             this.rowPos++;
@@ -152,11 +137,8 @@ export class Brick {
             this.handleLanded();
             this.isCurrentBrickLanded = true;
             this.game.generateNewBrick();
-            // this.game.generateNextBrick();
         }
-
     }
-
 
     rotate() {
         if (this.isLanded) {
@@ -169,23 +151,20 @@ export class Brick {
             this.rotateSound();
         }
     }
-    rotateSound() {//âm thanh xoay khối gạch 
+    rotateSound() {
         const audio = new Audio('../assets/audio/rotate.mp3');
         audio.play();
     }
 
-    fallFastSound() {//âm thanh khối gạch rơi xuống
+    fallFastSound() {
         const audio = new Audio('../assets/audio/263006__dermotte__giant-step-1.mp3');
         audio.play();
     }
-    fallBlockSound() {//âm thanh khối gạch rơi xuống
+    fallBlockSound() {
         const audio = new Audio('../assets/audio/263006__dermotte__giant-step-1.mp3');
         audio.play();
     }
 
-
-
-    // Collision handling
     checkCollision(nextRow: number, nextCol: number, nextLayout: any) {
         for (let row = 0; row < nextLayout.length; row++) {
             for (let col = 0; col < nextLayout[row].length; col++) {
@@ -212,7 +191,7 @@ export class Brick {
 
     handleLanded() {
         if (this.rowPos <= 0) {
-            this.handleGameOver();
+            this.game.handleGameOver();
             return;
         }
         if (!this.isCurrentBrickLanded) {
@@ -223,20 +202,12 @@ export class Brick {
                     }
                 }
             }
-
             this.board.handleCompletRows();
             this.board.drawBoard();
         }
     }
-    handleGameOver() {
-        this.gameOver = true;
-        this.game.showGameOverScreen();
-        this.gameoversound();
-    }
-    gameoversound() {//âm thanh khối gạch rơi xuống
+    gameoversound() {
         const audio = new Audio('../assets/audio/gameover.mp3');
         audio.play();
     }
-
-
 }
